push facility user dashboard work

+++ lib/models/facility_user_models/facilityModel/facility_model.dart
@@ -15,8 +15,9 @@ class FacilityModel {
   String? city;
   String? stateName;
   int? emrId;
-  int? monthlyCharge;
+  double? monthlyCharge;
   int? organizationId;
+
 
   FacilityModel(
       {this.id,
